+++ common/app/views/support/JavaScriptPage.scala
@@ -10,6 +10,16 @@ import conf.{Configuration, DiscussionAsset}
 import model._
 import play.api.libs.json._
 import play.api.mvc.RequestHeader
+import model.PageWithStoryPackage
+
+case class dcrJavaScriptPageConfigurationFragment(
+  hasShowcaseMainElement: Boolean,
+  isFront: Boolean,
+  isLiveblog: Boolean,
+  isMinuteArticle: Boolean,
+  isPaidContent: Boolean,
+  isSensitive: Boolean
+)
 
 object JavaScriptPage {
 
@@ -85,4 +95,16 @@ object JavaScriptPage {
       ("discussionFrontendUrl", JsString(DiscussionAsset("discussion-frontend.preact.iife")))
     )
   }
+
+  def commercialConfigurationFragmentForDotcomRendering(articlePage: PageWithStoryPackage, request: RequestHeader): dcrJavaScriptPageConfigurationFragment = {
+    dcrJavaScriptPageConfigurationFragment(
+      getMap(articlePage, Edition(request), false).getOrElse("hasShowcaseMainElement", JsBoolean(false)).as[Boolean],
+      getMap(articlePage, Edition(request), false).getOrElse("isFront", JsBoolean(false)).as[Boolean],
+      getMap(articlePage, Edition(request), false).getOrElse("isLiveBlog", JsBoolean(false)).as[Boolean],
+      getMap(articlePage, Edition(request), false).getOrElse("isMinuteArticle", JsBoolean(false)).as[Boolean],
+      getMap(articlePage, Edition(request), false).getOrElse("isPaidContent", JsBoolean(false)).as[Boolean],
+      getMap(articlePage, Edition(request), false).getOrElse("isSensitive", JsBoolean(false)).as[Boolean]
+    )
+  }
+
 }
